Fix position of navbar links
The position of the second item in the navbar was a little off, which
was clearly visible when hovering over the `Contribute` link. Setting
all `li` elements in the navbar to float to the left seems to fix the
issue on all viewports. This change required me to also add an extra
class `.border-left` to the first nav item `About`, which can be seen
in the previous commit.