Fix navbar item margin differently
Before, I used a `float` with value `left` to fix the navbar spacing as
described before. However, this messes up the navbar’s border. Another
way to address the issue is by setting a negative margin of `-4px` on
the left side of each (except the first) `nav li` item. The added class
`.border-left` is not necessary anymore.