Clarify the interaction of the CC-BY license and trademark rights
This commit changes the 'but does not grant you any trademark permissions' to a parenthetical comment
(an 'aside'), instead of having be part of the main sentence. With it part of the sentence, the reader
may believe that 'so long as you note the license and give credit' applies to the trademark-related
phrase, when in fact it relates the prior sections of the sentence.