Swap out Rust CoC suggestion with Citizen CoC
The Rust CoC defers to the Citizen Code of Conduct to define what harassment is. This deferment is without revision history. Instead the Citizen Code of Conduct should be presented as example CoC instead as it is self contained and can be referenced by revision to ensure that future changes are not snuck into many projects CoC.