Remove 'Beyond' section from doc/roadmap.md
Since the guides have been released, the Beyond section is no longer
relevant, so I am removing it.

"github.com/open-source is a landing page for a bunch of open source
things, but AFAIK right now we plan to keep the Open Source Guides on
opensource.guide. We will link to the Guides from github.com/open-source
though." - @nayafia