Move test dependencies, stop ignoring `node_modules`
Previously, `node_modules` was in `.gitignore` because there was a lot
of dependencies for the tests that don't need to be committed. However,
npm is also used to manage the primer dependency, which _does_ need to
be committed so GitHub Pages can build the website. At some point,
someone (whose name rhymes with "Landon Creepers") thought it'd be fine
to just type `git add -f node_modules/primer*` to commit the primer
dependencies and leave everything else ignored. This works…until you
want to update primer. Any update requires that the author knows about
this awful `.gitignore` setup, and uses the dreadful `-f` flag.

This commit:

- Moves the dependencies needed for tests into `test/package.json`
- Adds `test/node_modules` to `.gitignore`
- Removes `node_modules` from `.gitignore`