Add spell checking for articles
This commit adds spell checking to all the other prose and
markdown checks currently in place.
As spell checking is a pretty pedantic issue, often marking
jargon (like 501c3, BDFL, or grantmaking) and project and
product names (like Homebrew, Kubernetes, or WordPress),
a personal `dictionary.txt` is included in `test/` with
exceptions for the articles.

URLs, filepaths, @mentions, emoji, and text in block quotes
(as those are used for quotes in the articles) are excluded
from the spell checking.

The spell checking highlighted some cases that I changed:

*   All alt text of images is now properly cases if they
    start the paragraph
*   Typo `sizeable` > `sizable` (former is GB English, latter is
    US English)
*   Casing of `reddit` > `Reddit` (latter seems to be used more)
*   Casing of `pypa` > `PyPA` (latter seems to be used more)
*   Casing of `eslint` > `ESLint` (latter is the official casing)
*   Casing of `READMES` > `READMEs` (plural `s` is now cased just
    like `PRs` and `APIs`)
*   Casing of `ok` > `OK` (latter is the propper spelling)
*   Spelling of `PostgreSQL` > `Postgres` (both were used throughout,
    the latter seems simpler and is just as well known)
*   Hyphen in `tradeoffs` > `trade-offs` (both could probably be used,
    but the latter is more often used)
*   Latin `de minimis` > `minimal` (this section is about law, so the
    Latin makes sense, but it’s also a phrase I had to look up, so
    I’m guessing other people would have to as well)
*   Space in `pageviews` > `page views` (former is more about the
    metrics in Google Analytics I believe, latter is just as good?)
*   Space in `datasets` > `data sets` (both work, but latter is more
    common?)