Change translations approach
Instead of having each of the translations live on a separate GitHub
Pages site that we link out to let’s get all of the translations
inside this repository. This will make it easier for others to track
and modify all translations as well as making changes that affect all
translations in a single pass.