Translation Error Changes
My mother tongue is Chinese, and I found some errors and inappropriate expressions when I read this document. I hope that these changes I made can be adopted by you.
Also, this is my first contribution as a newcomer in the open-source world. Contributing to a document that teaches people how to contribute is awesome!