The rule of law
I was born in the city of Georgia, Tbilisi, and the reason I wan't to mention that is, because I used play a lot at internet cafe's back in a day, when nobody had computers in my country. Not only that, but on top of that electricity came on only for 2 hours a day. A lot of people I was friends with used to go to the same internet cafe as me and living in the city made that 10 times better. We used to have a blast, going out to play LAN almost every day. What I want to say is that, I lived in such a small country with no possibility of playing a game on international servers. Everybody we used to play with literally lived in a same city 10-15 minutes away,, so we had a lot of respect for each other. I was lucky to be in that situation, because, unfortunately, some people don't think what their words might do to somebody emotionally. Everybody has to remember that people are different, we got to understand that literally there is no duplicate personalities, and if there is it's probably just a 1% in the whole world. That means everybody's views, ideas, and thoughts, are different right?.  We have to accept everybody as they are. Maybe you are getting annoyed? Okay there is nobody invading your space right? There is nobody yelling at your face, so just be a decent human being and move on, it's as simple as that. So if you are the type of individual, to mock on people and intentionally try to make their day miserable in this day and age? You do not deserve to access internet, and should be removed ASAP.