Improve site method readability
Add documentation to site-related methods

This commit adds internal documentation to the methods in the Helper module that relate to the Jekyll site. The documentation explains what each method does and how it should be used. The code is also formatted according to the Ruby style guide for improved readability. These changes should make it easier for other developers to understand and work with the site-related code.