Add link to git-authorship in legal.md
Git Authorship is a project that helps maintainers identify the
individuals who have committed code present in the current version
of the project.

Added with the permission of @ahpook to avoid "self-promotion"
concerns since this is one of my own projects.

The commented section at the end of the article discusses ideas
for future additions to this article.