Remove references to the Personal Ecology doc, which has 404d
This document went away and tests fail because it is now 404ing. There is at least one other resource in both the inline text and the list context so it should be safe to simply remove the reference to it.

Note I may have introduced grammatical errors in the non-Roman character languages, because I couldn't tell whether the remaining words reference the doc which was deleted.