Update legal.md with archived link to jquery CLA
The original document has gone 404 as jquery moved to the js foundation and then openjsf. They also now use the Apache CLA via Linux Foundation's EasyCLA, so they are no longer a good example of a "lightweight" CLA project. It seemed most straightforward to replace this link with the archive.org equivalent, and just put our faith in the Internet Archive.