Remove version lock from html-proofer
There have been 2 major revs of HTML-Proofer since the one
we pinned. The current one is causing flaky tests - it's missing
clear errors w/internal links and only reports the first
error it finds rather than running to completion and reporting
all of them.

The latest version (5.0.8 at the moment) does catch all of the
errors but has a problem locally where UTF-8 encoded anchor links
are not properly decoded. This might be a problem local to me though;
I coded up a spec test in html-proofer with an exact copy-paste
of one of the failing links and ... it passes there.

So i'll unpin this and see if there's some local gem issue that's
causing this on my machine and maybe it works in the Actions env.